refactor: utils에 index.ts 추가
